Use local month rather than UTC Month to set date
We are attempting to find "The first of the month following this one" in the _user_'s timezone. By setting the timestamp 8 hours into the day, we will always make sure that the subscription refresh time is in the future, even if someone signs up at 11:59pm in Mountain View on the last day of the month.