Combines site and building code into one field
Since site code can be derived from building code, combines the fields into a single input for building code, and the site code field for our data is extracted from the building code upon submission. If the building code provided is not in proper building code format, the site code is set as whatever is provided in the building code.